+++ joern-cli/frontends/gosrc2cpg/src/main/scala/io/joern/gosrc2cpg/astcreation/AstCreator.scala
@@ -25,6 +25,7 @@ class AstCreator(val relPathFileName: String, val parserResult: ParserResult)(im
     with AstForFunctionsCreator
     with AstForPrimitivesCreator
     with AstForStatementsCreator
+    with AstForStructuresCreator
     with X2CpgAstNodeBuilder[ParserNodeInfo, AstCreator] {
 
   protected val logger: Logger = LoggerFactory.getLogger(classOf[AstCreator])
