+++ joern-cli/frontends/gosrc2cpg/src/main/scala/io/joern/gosrc2cpg/astcreation/AstForGenDeclarationCreator.scala
@@ -27,41 +27,6 @@ trait AstForGenDeclarationCreator(implicit withSchemaValidation: ValidationMode)
       .toSeq
   }
 
-  private def astForTypeSpec(typeSpecNode: ParserNodeInfo): Seq[Ast] = {
-    // TODO: Add support for member variables and methods
-    Option(typeSpecNode) match {
-      case Some(typeSpec) =>
-        val nameNode          = typeSpec.json(ParserKeys.Name)
-        val typeNode          = createParserNodeInfo(typeSpec.json(ParserKeys.Type))
-        val a: NewTypeDecl    = methodAstParentStack.collectFirst { case t: NewTypeDecl => t }.get
-        val astParentType     = a.label
-        val astParentFullName = a.fullName
-        val fullName          = nameNode(ParserKeys.Name).str // TODO: Discuss fullName structure
-        val typeDeclNode_ =
-          typeDeclNode(
-            typeSpec,
-            nameNode(ParserKeys.Name).str,
-            fullName,
-            parserResult.filename,
-            typeSpec.code,
-            astParentType,
-            astParentFullName
-          )
-
-        methodAstParentStack.push(typeDeclNode_)
-        scope.pushNewScope(typeDeclNode_)
-        val memberAsts = astForStructType(typeNode)
-
-        methodAstParentStack.pop()
-        scope.popScope()
-
-        val modifier = addModifier(typeDeclNode_, nameNode(ParserKeys.Name).str)
-        Seq(Ast(typeDeclNode_).withChild(Ast(modifier)).withChildren(memberAsts.toSeq))
-      case None =>
-        Seq.empty
-    }
-  }
-
   private def astForImport(nodeInfo: ParserNodeInfo): Seq[Ast] = {
     val basicLit       = createParserNodeInfo(nodeInfo.json(ParserKeys.Path))
     val importedEntity = nodeInfo.json(ParserKeys.Path).obj(ParserKeys.Value).str.replaceAll("\"", "")
